requirements.txt
openai>=1.0.0
pdfplumber>=0.11.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
tqdm>=4.65.0
markdown>=3.5.0
EbookLib>=0.18
beautifulsoup4>=4.12.0
